Read Octopus Agile Daily Prices
This module reads the prices published daily for the upcoming costs of electricity in 30 min slots.
It extracts the Peak and Minimum charging times, and sets flags to indicate when the periods have started. These can then be used to trigger scenes, e.g. to turn on car charging overnight, increase house temp prior to peak, and reduce it during peak hours.